Moved files to appropriate folders
