+++ core/src/com/mygdx/game/controller/FirebaseInterface.java
@@ -1,4 +1,6 @@
-package com.mygdx.game;
+package com.mygdx.game.controller;
+
+import com.mygdx.game.model.components.Score;
 
 import java.util.ArrayList;
