+++ core/src/com/mygdx/game/model/components/Score.java
@@ -1,4 +1,4 @@
-package com.mygdx.game;
+package com.mygdx.game.model.components;
 
 public class Score {
